trustfence: remove compiler warning with appropriate casting
 warning: cast to pointer from integer of different size [-Wint-to-pointer-cast]
    uint8_t *dek_blob = (uint8_t*)(csf_addr + CONFIG_CSF_SIZE - blob_size);

Manual cherry-pick from b5af89a9008372fc79b10cb801f80087933e51b2.

+++ cmd/trustfence.c
@@ -75,7 +75,8 @@ static int get_dek_blob(char *output, u32 *size) {
 
 	if (*csf_addr) {
 		int blob_size = MAX_DEK_BLOB_SIZE;
-		uint8_t *dek_blob = (uint8_t *)(*csf_addr + CONFIG_CSF_SIZE - blob_size);
+		uint8_t *dek_blob = (uint8_t *)(uintptr_t)(*csf_addr +
+				    CONFIG_CSF_SIZE - blob_size);
 
 		/*
 		 * Several DEK sizes can be used.
